pyproject.toml
[tool.poetry]
name = "19-dateninfrastruktur"
version = "0.1.0"
description = ""
authors = ["m.habedank <m.habedank@cartelsol.com>"]
readme = "README.md"
packages = [{include = "19_dateninfrastruktur"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12,"
boto3 = "^1.28.15"
notebook = "^7.0.0"
python-dotenv = "^1.0.0"
pandas = "^2.0.3"
sqlalchemy = "^2.0.19"
psycopg2-binary = "^2.9.6"
sqlalchemy-utils = "^0.41.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
peewee = "^3.16.2"
pandera = "^0.16.1"
pytest = "^7.4.0"
setuptools = "^68.0.0"
click = "^8.1.6"
black = "^23.7.0"
fastparquet = "^2023.7.0"
pyarrow = "^12.0.1"
shapely = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
